ci: add Dockerfile (#31)
* ci: add uv-export to pre-commit to export requirements.txt

* ci: run ruff format on missed file

* ci: apply consistent naming to workflows

* ci: add Dockerfile and ci build workflow

* ci: remove hashes from requirements.txt

caused by this issue https://github.com/pypa/pip/issues/4995

* ci: add pyproject.toml

build fails without it:

ERROR: file:///app (from -r requirements.txt (line 3)) does not appear
to be a Python project: neither 'setup.py' nor 'pyproject.toml' found.

* ci: switch to non-editable dependencies to make Docker build easier

otherwise would require Hatch build in the Docker build

* ci: remove fluff from pyproject.toml

* ci: add .dockerignore and use non-root user for runtime

* refactor: fix incorrect migration

found when switching to Postgres. media_format should have the shorthand
value stored in the database.

* ci: add Postgres support in production + extract env variables

* fix: squash initial migrations to bypass date to int type casting issues with Postgres

migration 0008 casts from a date type to int. this fails in Postgres, so
I've squashed all migrations together up till 0008 and changed the
original definition of the field to be an integer field to prevent the
type casting from happening.

* ci: flesh out Dockerfile

* ci: add compose.yml with Caddy and Postgres

* refactor: split dev and prod settings files

* ci: add development stage to Dockerfile

* ci: split compose environments + create test env that uses SQLite

* ci: add Docker Compose Watch to dev environment

* ci: move Docker assets into docker folder

cleans up the codebase a little bit and allows for a more organised
structure and infrastructure tests

* test: randomly generate a SECRET_KEY for tests

this is a required parameter, and Django will fail to start in the tests
if we don't provide a SECRET_KEY

* ci: fix incorrect Dockerfile paths in build workflow

* ci: set target stage to production

* test: fix failing tests

* test: add Docker image infra tests

* ci: install all dependencies in CI

* test: fix incorrect image tag

* test: fix incorrect image tag

* ci: ignore docker tests from regular test suite

* test: use builtin host.pip.get_packages() method for getting dependencies

* test: improve and parametrize test suite

* test: switch to using dgoss for container tests + use compose for CI builds

reduce manual implementation effort and rely on existing tools. make CI
builds the same as local builds.

* ci: remove non-existent field

* ci: fix incorrect context path

* ci: remove test packages

* ci: use local source rather than remote git context

* ci: path issues

* ci: path issues

* ci: path issues

* ci: path issues

* ci: switch back to build-push-action

Bake Action can't resolve the compose.core.yml reference in
compose.prod.yml. it's expecting to find it in the context directory,
which is the root directory.

* ci: set Docker test working directory

* ci: remove unnecessary ignore