refactor: convert all movie import tests to async
these don't actually need to be async because the manage.py command
would get run in a separate process to the webserver and won't block
anything. the manage.py BaseCommand is also sync only. I've just
converted these to async so there's general consistency across the app
and the way the ORM is used.

also configured typos to not change objects.acount spelling.