feat: switch to async (#34)
* feat: add Gunicorn for running Uvicorn workers

performance improvement as per recommendations in Uvicorn docs

* feat: convert application to async - WIP

* refactor: add django-types + type hints to models

* refactor: add explicit ForeignKey releate_names

* refactor: convert model methods to async where applicable

* refactor: convert all API tests to async

* refactor: convert all movie import tests to async

these don't actually need to be async because the manage.py command
would get run in a separate process to the webserver and won't block
anything. the manage.py BaseCommand is also sync only. I've just
converted these to async so there's general consistency across the app
and the way the ORM is used.

also configured typos to not change objects.acount spelling.

* ci: move pyright typeCheckingMode setting into pyproject.toml

* refactor: use correct related names and RelatedManager type

this gets most of the type safety I need

* refactor: fix RelatedManager runtime issue

* test: move fixture into conftest.py

* test: use Protocol instead of Callable for type hints + create Bookcase factory fixture

Protocol seems to be the recommended approach to type hinting a callable
with multiple runtime signatures (optional/default arguments)

* test: implement MediaCaseDimensions factory fixture + convert tests to async

* test: implement Collection factory fixture + convert tests to use it

tests are in a broken state because there's a mix of sync and async with
the Collection tests. this will be fixed when the rest of the models are
converted to use the async factory fixtures.

* test: migrate Shelf tests to async

test: migrate all Shelf tests to async + improve factory fixtures

allow fields that we don't normally care about to be left out in factory
arguments

* refactor: add default ordering to Movie class

* refactor: nearly finished moving to async

* ci: add test report to CI

* test: finish moving tests over to async

wrapped some baker.make calls in sync to async to cut down on
boilerplate and complexity with having a bunch of fixture factories

* refactor: move __str__ methods to __repr__ and implement new __str__ methods

the existing __str__ methods were more like __repr__. tests adapted to
only test the __str__ method resolution.

* test: use transactions for db tests

* test: move all tests into tests/ directory

* test: get tests passing

there's some issues with pytest asyncio and properly rolling back the
database state. will work on improving this.

* refactor: make release_year optional in make_movie fixture factory

cut down on boilerplate

* refactor: fix failing test - add ordering to Movie model