feat: improve main api endpoints (#53)
* refactor: ignore type errors caused by poor library type hinting

* feat: add ShelfDimension endpoints

* feat: add more endpoints and schema

* refactor: do pagination by default

* ci: add pyright type stubs to project for Django

add custom aget_object_or_404 to solve issues

* refactor: resolve pyright issues by using type stubs

* ci: add pyright type stubs for django-ninja

* refactor: namespace models in schema.py

so we can cut down on the verbose naming of classes and call the schemas
MovieIn and MovieOut instead of MovieSchemaIn and MovieSchemaOut and not
have to worry about name clashes

* refactor: namespace schemas in api.py

so we can rename all schemas to follow the simpler format of MovieIn and
MovieOut

* refactor: rename schemas to follow simplified naming convention

don't need to include Schema in the name as they're already namespaced
in schemas.py

* refactor: make MediaCaseDimension singular instead of plural

to make naming consistent

* refactor: make ShelfDimension singular instead of plural

* refactor: move MediaFormat TextChoices inside dependent model

* refactor: add help text and additional validation to Dimension class

* refactor: rename PhysicalMediaOrientation to ShelfOrientation and move inside dependent class

* refactor: shorten name to Orientation since it's now namespaced

* refactor: shorten name to Format since it's now namespaced

* refactor: only apply unique constraint to physical media position on shelf when position is not None

this check ensures that we only enforce the validation when media is
actually assigned to a shelf

* refactor: simplify mapping Shelf.Orientation values to "height" and "width" Dimension properties

* refactor: rename case_dimensions to dimensions on PhysicalMedia

it's already namespaced, so don't need to include "case_". this brings
the API in line with the Shelf.dimension property.

* refactor: get rid of getattr in favour of helper method

* refactor: adapt schemas for in/out to inherit from base schema

* refactor: split up endpoints into separate files

* test: fix broken tests caused by 302 redirect to include pagination

pagination is enforced on list endpoints

* test: fix test errors caused by random generation of Dimension.depth values

introduction of the depth field has meant that bake now randomly
generates a value for these fields. media depth must always be less than
shelf depth to satisfy the can_fit_media method